incomplete: remove all services from processManagers. ToDo: pass relevant data to visualization and remove all redundant services
