+++ src/main/kotlin/org/jetbrains/research/testspark/services/ClearService.kt
@@ -1,9 +1,6 @@
 package org.jetbrains.research.testspark.services
 
 import com.intellij.openapi.components.Service
-import com.intellij.openapi.components.service
-import com.intellij.openapi.project.Project
-import org.jetbrains.research.testspark.core.utils.DataFilesUtil
 
 @Service(Service.Level.PROJECT)
 class ClearService {
@@ -12,13 +9,13 @@ class ClearService {
      * error service, coverage visualization, and test generation data.
      *
      * @param project the project to clear the test-related data for
-     */
-    fun clear(project: Project) {
-        project.service<TestCaseDisplayService>().clear()
-        project.service<ErrorService>().clear()
-        project.service<CoverageVisualisationService>().clear()
-        project.service<TestGenerationDataService>().clear()
-        project.service<TestsExecutionResultService>().clear()
-        DataFilesUtil.cleanFolder(project.service<ProjectContextService>().resultPath!!)
-    }
+//     */
+//    fun clear(project: Project) { // should be removed totally!
+//        project.service<TestCaseDisplayService>().clear()
+//        project.service<ErrorService>().clear()
+//        project.service<CoverageVisualisationService>().clear()
+//        project.service<TestGenerationData>().clear()
+//        project.service<TestsExecutionResultService>().clear()
+//        DataFilesUtil.cleanFolder(project.service<ProjectContextService>().resultPath!!)
+//    }
 }
