+++ src/main/kotlin/org/jetbrains/research/testspark/services/LLMChatService.kt
@@ -2,13 +2,11 @@ package org.jetbrains.research.testspark.services
 
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.project.Project
-import org.jetbrains.research.testspark.bundles.TestSparkBundle
-import org.jetbrains.research.testspark.core.generation.llm.network.LLMResponse
 import org.jetbrains.research.testspark.core.progress.CustomProgressIndicator
-import org.jetbrains.research.testspark.tools.llm.SettingsArguments
-import org.jetbrains.research.testspark.tools.llm.error.LLMErrorManager
-import org.jetbrains.research.testspark.tools.llm.generation.StandardRequestManagerFactory
 import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
+import org.jetbrains.research.testspark.tools.llm.SettingsArguments
+import org.jetbrains.research.testspark.tools.llm.generation.StandardRequestManagerFactory
+import org.jetbrains.research.testspark.tools.llm.isCorrectToken
 
 @Service(Service.Level.PROJECT)
 class LLMChatService(private val project: Project) {
@@ -18,9 +16,9 @@ class LLMChatService(private val project: Project) {
     /**
      * Re-initiates the requestManager. All the chat history will be removed.
      */
-    fun newSession() {
-        requestManager = StandardRequestManagerFactory().getRequestManager(project)
-    }
+//    fun newSession() {
+//        requestManager = StandardRequestManagerFactory().getRequestManager(project)
+//    }
 
     /**
      * Sends a test generation request (when chat history of the requestManager is empty).
@@ -37,13 +35,13 @@ class LLMChatService(private val project: Project) {
      * the TestSuiteGeneratedByLLM component represents the test suite parsed from the LLm response.
      * If the test suite generation fails, the TestSuiteGeneratedByLLM object will be null. and the reason is available in the string.
      */
-    fun testGenerationRequest(
-        messageToPrompt: String,
-        indicator: CustomProgressIndicator,
-        packageName: String,
-    ): LLMResponse {
-        return requestManager.request(messageToPrompt, indicator, packageName)
-    }
+//    fun testGenerationRequest(
+//        messageToPrompt: String,
+//        indicator: CustomProgressIndicator,
+//        packageName: String,
+//    ): LLMResponse {
+//        return requestManager.request(messageToPrompt, indicator, packageName)
+//    }
 
     /**
      * Sends a test modification request according to user's feedback.
@@ -97,19 +95,8 @@ class LLMChatService(private val project: Project) {
      */
     private fun updateToken(): Boolean {
         requestManager.token = SettingsArguments.getToken()
-        return isCorrectToken()
+        return isCorrectToken(project)
     }
 
-    /**
-     * Checks if the token is set.
-     *
-     * @return True if the token is set, false otherwise.
-     */
-    fun isCorrectToken(): Boolean {
-        if (!SettingsArguments.isTokenSet()) {
-            LLMErrorManager().errorProcess(TestSparkBundle.message("missingToken"), project)
-            return false
-        }
-        return true
-    }
+
 }
