+++ src/main/kotlin/org/jetbrains/research/testspark/tools/ToolUtils.kt
@@ -1,6 +1,7 @@
 package org.jetbrains.research.testspark.tools
 
 import com.intellij.openapi.components.service
+import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.module.ModuleManager
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.CompilerModuleExtension
@@ -10,12 +11,12 @@ import org.jetbrains.research.testspark.core.utils.DataFilesUtil
 import org.jetbrains.research.testspark.data.Report
 import org.jetbrains.research.testspark.services.ErrorService
 import org.jetbrains.research.testspark.services.JavaClassBuilderService
-import org.jetbrains.research.testspark.services.ProjectContextService
-import org.jetbrains.research.testspark.services.TestGenerationDataService
-import org.jetbrains.research.testspark.services.TestStorageProcessingService
+import org.jetbrains.research.testspark.services.TestGenerationData
 import org.jetbrains.research.testspark.services.TestsExecutionResultService
+import org.jetbrains.research.testspark.tools.llm.getClassWithTestCaseName
 import java.io.File
 
+private val log = Logger.getInstance("TestSparkToolUtils")
 /**
  * Retrieves the imports code from a given test suite code.
  *
@@ -61,27 +62,29 @@ fun saveData(
     report: Report,
     packageLine: String,
     importsCode: MutableSet<String>,
+    fileUrl: String,
+    generatedTestData: TestGenerationData
 ) {
-    project.service<TestGenerationDataService>().resultName = project.service<TestStorageProcessingService>().testResultName
-    project.service<TestGenerationDataService>().fileUrl = project.service<ProjectContextService>().fileUrl!!
-    project.service<TestGenerationDataService>().packageLine = packageLine
-    project.service<TestGenerationDataService>().importsCode.addAll(importsCode)
+    generatedTestData.fileUrl = fileUrl
+    generatedTestData.packageLine = packageLine
+    generatedTestData.importsCode.addAll(importsCode)
 
     project.service<TestsExecutionResultService>().initExecutionResult(report.testCaseList.values.map { it.id })
 
     for (testCase in report.testCaseList.values) {
         val code = testCase.testCode
         testCase.testCode = project.service<JavaClassBuilderService>().generateCode(
-            project.service<JavaClassBuilderService>().getClassWithTestCaseName(testCase.testName),
+            getClassWithTestCaseName(testCase.testName),
             code,
-            project.service<TestGenerationDataService>().importsCode,
-            project.service<TestGenerationDataService>().packageLine,
-            project.service<TestGenerationDataService>().runWith,
-            project.service<TestGenerationDataService>().otherInfo,
+            generatedTestData.importsCode,
+            generatedTestData.packageLine,
+            generatedTestData.runWith,
+            generatedTestData.otherInfo,
+            generatedTestData
         )
     }
 
-    project.service<TestGenerationDataService>().testGenerationResultList.add(report)
+    project.service<TestGenerationData>().testGenerationResultList.add(report)
 }
 
 /**
@@ -140,3 +143,5 @@ fun processStopped(project: Project, indicator: CustomProgressIndicator): Boolea
     }
     return false
 }
+
+
