+++ src/main/kotlin/org/jetbrains/research/testspark/tools/generatedTests/TestProcessor.kt
@@ -1,0 +1,4 @@
+package org.jetbrains.research.testspark.tools.generatedTests
+
+class TestProcessor {
+}
