+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/generation/LLMProcessManager.kt
@@ -4,28 +4,27 @@ import com.intellij.openapi.components.service
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
 import org.jetbrains.research.testspark.bundles.TestSparkBundle
+import org.jetbrains.research.testspark.bundles.TestSparkToolTipsBundle
 import org.jetbrains.research.testspark.core.generation.llm.network.LLMResponse
 import org.jetbrains.research.testspark.core.generation.llm.network.ResponseErrorCode
 import org.jetbrains.research.testspark.core.progress.CustomProgressIndicator
-import org.jetbrains.research.testspark.bundles.TestSparkToolTipsBundle
+import org.jetbrains.research.testspark.core.test.data.TestCaseGeneratedByLLM
+import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
 import org.jetbrains.research.testspark.data.FragmentToTestData
+import org.jetbrains.research.testspark.data.ProjectContext
 import org.jetbrains.research.testspark.data.Report
 import org.jetbrains.research.testspark.data.TestCase
 import org.jetbrains.research.testspark.services.ErrorService
-import org.jetbrains.research.testspark.services.JavaClassBuilderService
-import org.jetbrains.research.testspark.services.LLMChatService
-import org.jetbrains.research.testspark.services.ProjectContextService
 import org.jetbrains.research.testspark.services.SettingsProjectService
-import org.jetbrains.research.testspark.services.TestGenerationDataService
-import org.jetbrains.research.testspark.services.TestStorageProcessingService
+import org.jetbrains.research.testspark.services.TestGenerationData
+import org.jetbrains.research.testspark.tools.generatedTests.TestUtils
 import org.jetbrains.research.testspark.tools.getBuildPath
 import org.jetbrains.research.testspark.tools.getImportsCodeFromTestSuiteCode
 import org.jetbrains.research.testspark.tools.getPackageFromTestSuiteCode
 import org.jetbrains.research.testspark.tools.llm.SettingsArguments
 import org.jetbrains.research.testspark.tools.llm.error.LLMErrorManager
-import org.jetbrains.research.testspark.core.test.data.TestCaseGeneratedByLLM
+import org.jetbrains.research.testspark.tools.llm.getClassWithTestCaseName
 import org.jetbrains.research.testspark.tools.llm.test.TestSuitePresenter
-import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
 import org.jetbrains.research.testspark.tools.processStopped
 import org.jetbrains.research.testspark.tools.saveData
 import org.jetbrains.research.testspark.tools.template.generation.ProcessManager
@@ -47,11 +46,11 @@ class LLMProcessManager(
     private val promptManager: PromptManager,
     private val testSamplesCode: String,
 ) : ProcessManager {
-    private val settingsProjectState = project.service<SettingsProjectService>().state
     private val testFileName: String = "GeneratedTest.java"
     private val log = Logger.getInstance(this::class.java)
     private val llmErrorManager: LLMErrorManager = LLMErrorManager()
     private val maxRequests = SettingsArguments.maxLLMRequest()
+    private val testUtils = TestUtils(project)
 
     /**
      * Runs the test generator process.
@@ -64,14 +63,16 @@ class LLMProcessManager(
         indicator: CustomProgressIndicator,
         codeType: FragmentToTestData,
         packageName: String,
+        projectContext: ProjectContext,
+        generatedTestsData: TestGenerationData
     ) {
         log.info("LLM test generation begins")
 
         if (processStopped(project, indicator)) return
 
         // update build path
-        var buildPath = project.service<ProjectContextService>().projectClassPath!!
-        if (settingsProjectState.buildPath.isEmpty()) {
+        var buildPath = projectContext.projectClassPath!!
+        if (project.service<SettingsProjectService>().state.buildPath.isEmpty()) {
             // User did not set own path
             buildPath = getBuildPath(project)
         }
@@ -93,9 +94,8 @@ class LLMProcessManager(
         var messageToPrompt = promptManager.generatePrompt(codeType, testSamplesCode)
         var generatedTestSuite: TestSuiteGeneratedByLLM? = null
 
-        // notify LLMChatService to restart the chat process.
-        project.service<LLMChatService>().newSession()
-
+        // Initiate a new RequestManager
+        val requestManager = StandardRequestManagerFactory().getRequestManager(project)
         // Asking LLM to generate test. Here, we have a loop to make feedback cycle for LLm in case of wrong responses.
         while (!generatedTestsArePassing) {
             requestsCount++
@@ -108,7 +108,7 @@ class LLMProcessManager(
             }
 
             // Ending loop checking
-            if (isLastIteration(requestsCount) && project.service<TestGenerationDataService>().compilableTestCases.isEmpty()) {
+            if (isLastIteration(requestsCount) && generatedTestsData.compilableTestCases.isEmpty()) {
                 llmErrorManager.errorProcess(TestSparkBundle.message("invalidLLMResult"), project)
                 break
             }
@@ -119,8 +119,7 @@ class LLMProcessManager(
             }
 
             val response: LLMResponse =
-                project.service<LLMChatService>().testGenerationRequest(messageToPrompt, indicator, packageName)
-
+                requestManager.request(messageToPrompt, indicator, packageName)
             when(response.errorCode) {
                 ResponseErrorCode.OK -> {
                     log.info("Test suite generated successfully: ${response.testSuite!!}")
@@ -163,22 +162,22 @@ class LLMProcessManager(
 
             // Save the generated TestSuite into a temp file
             val generatedTestCasesPaths: MutableList<String> = mutableListOf()
-            val testSuitePresenter = TestSuitePresenter(project)
+            val testSuitePresenter = TestSuitePresenter(project, generatedTestsData)
 
             if (isLastIteration(requestsCount)) {
-                generatedTestSuite.updateTestCases(project.service<TestGenerationDataService>().compilableTestCases.toMutableList())
+                generatedTestSuite.updateTestCases(generatedTestsData.compilableTestCases.toMutableList())
             }
             else {
                 for (testCaseIndex in generatedTestSuite.testCases.indices) {
-                    val testFileName = "${project.service<JavaClassBuilderService>().getClassWithTestCaseName(generatedTestSuite.testCases[testCaseIndex].name)}.java"
+                    val testFileName = "${getClassWithTestCaseName(generatedTestSuite.testCases[testCaseIndex].name)}.java"
 
                     val testCaseRepresentation = testSuitePresenter
                         .toStringSingleTestCaseWithoutExpectedException(generatedTestSuite, testCaseIndex)
 
-                    val saveFilepath = project.service<TestStorageProcessingService>().saveGeneratedTest(
+                    val saveFilepath = testUtils.saveGeneratedTest(
                         generatedTestSuite.packageString,
                         testCaseRepresentation,
-                        project.service<ProjectContextService>().resultPath!!,
+                        generatedTestsData.resultPath,
                         testFileName
                     )
 
@@ -186,10 +185,10 @@ class LLMProcessManager(
                 }
             }
 
-            val generatedTestPath: String = project.service<TestStorageProcessingService>().saveGeneratedTest(
+            val generatedTestPath: String = testUtils.saveGeneratedTest(
                 generatedTestSuite.packageString,
                 testSuitePresenter.toStringWithoutExpectedException(generatedTestSuite),
-                project.service<ProjectContextService>().resultPath!!,
+                generatedTestsData.resultPath,
                 testFileName,
             )
 
@@ -206,14 +205,14 @@ class LLMProcessManager(
                 if (!isLastIteration(requestsCount)) {
                     generatedTestSuite.testCases
                 } else {
-                    project.service<TestGenerationDataService>().compilableTestCases.toMutableList()
+                    generatedTestsData.compilableTestCases.toMutableList()
                 }
 
             // Compile the test file
             indicator.setText(TestSparkBundle.message("compilationTestsChecking"))
 
-            val separateCompilationResult = project.service<TestStorageProcessingService>().compileTestCases(generatedTestCasesPaths, buildPath, testCases)
-            val commonCompilationResult = project.service<TestStorageProcessingService>().compileCode(File(generatedTestPath).absolutePath, buildPath)
+            val separateCompilationResult = testUtils.compileTestCases(generatedTestCasesPaths, buildPath, testCases, generatedTestsData)
+            val commonCompilationResult = testUtils.compileCode(File(generatedTestPath).absolutePath, buildPath)
 
             if (!separateCompilationResult && !isLastIteration(requestsCount)) {
                 log.info("Incorrect result: \n${testSuitePresenter.toString(generatedTestSuite)}")
@@ -238,7 +237,7 @@ class LLMProcessManager(
 
         log.info("Result is ready")
 
-        val testSuitePresenter = TestSuitePresenter(project)
+        val testSuitePresenter = TestSuitePresenter(project,generatedTestsData)
         val testSuiteRepresentation =
             if (generatedTestSuite != null) testSuitePresenter.toString(generatedTestSuite) else null
 
@@ -246,8 +245,10 @@ class LLMProcessManager(
             project,
             report,
             getPackageFromTestSuiteCode(testSuiteRepresentation/*generatedTestSuite.toString()*/),
-            getImportsCodeFromTestSuiteCode(testSuiteRepresentation/*generatedTestSuite.toString()*/, project.service<ProjectContextService>().classFQN!!),
-        )
+            getImportsCodeFromTestSuiteCode(testSuiteRepresentation/*generatedTestSuite.toString()*/, projectContext.classFQN!!,),
+            projectContext.fileUrl!!,
+            generatedTestsData
+            )
     }
 
     private fun isLastIteration(requestsCount: Int) = requestsCount > maxRequests
