+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/test/TestSuitePresenter.kt
@@ -4,9 +4,14 @@ import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import org.jetbrains.research.testspark.core.test.data.TestSuiteGeneratedByLLM
 import org.jetbrains.research.testspark.services.JavaClassBuilderService
+import org.jetbrains.research.testspark.services.TestGenerationData
+import org.jetbrains.research.testspark.tools.llm.getClassWithTestCaseName
 
 
-class TestSuitePresenter(private val project: Project) {
+class TestSuitePresenter(
+    private val project: Project,
+    val generatedTestsData: TestGenerationData
+) {
     /**
      * Returns a string representation of this object.
      *
@@ -38,6 +43,7 @@ class TestSuitePresenter(private val project: Project) {
                 packageString,
                 runWith,
                 otherInfo,
+                generatedTestsData
             )
         }
     }
@@ -52,12 +58,13 @@ class TestSuitePresenter(private val project: Project) {
         testSuite: TestSuiteGeneratedByLLM, testCaseIndex: Int): String =
         testSuite.run {
             project.service<JavaClassBuilderService>().generateCode(
-                project.service<JavaClassBuilderService>().getClassWithTestCaseName(testCases[testCaseIndex].name),
+                getClassWithTestCaseName(testCases[testCaseIndex].name),
                 testCases[testCaseIndex].toStringWithoutExpectedException() + "\n",
                 imports,
                 packageString,
                 runWith,
                 otherInfo,
+                generatedTestsData
             )
         }
 
@@ -80,6 +87,7 @@ class TestSuitePresenter(private val project: Project) {
                 packageString,
                 runWith,
                 otherInfo,
+                generatedTestsData
             )
         }
     }
