+++ src/main/kotlin/org/jetbrains/research/testspark/tools/template/generation/ProcessManager.kt
@@ -2,6 +2,8 @@ package org.jetbrains.research.testspark.tools.template.generation
 
 import org.jetbrains.research.testspark.core.progress.CustomProgressIndicator
 import org.jetbrains.research.testspark.data.FragmentToTestData
+import org.jetbrains.research.testspark.data.ProjectContext
+import org.jetbrains.research.testspark.services.TestGenerationData
 
 /**
  * An interface representing a process manager.
@@ -18,5 +20,7 @@ interface ProcessManager {
         indicator: CustomProgressIndicator,
         codeType: FragmentToTestData,
         packageName: String,
+        projectContext: ProjectContext,
+        generatedTestData: TestGenerationData
     )
 }
